+++ recorder/src/element.ts
@@ -1,42 +1,3 @@
-// --
-// from playwright to match their text engine
-function shouldSkipForTextMatching(element: Element | ShadowRoot) {
-  return (
-    element.nodeName === "SCRIPT" ||
-    element.nodeName === "STYLE" ||
-    (document.head && document.head.contains(element))
-  );
-}
-
-export function elementText(root: Element | ShadowRoot): string {
-  let value = "";
-
-  if (!shouldSkipForTextMatching(root)) {
-    if (
-      root instanceof HTMLInputElement &&
-      (root.type === "submit" || root.type === "button")
-    ) {
-      value = root.value;
-    } else {
-      for (let child = root.firstChild; child; child = child.nextSibling) {
-        if (child.nodeType === Node.ELEMENT_NODE)
-          value += elementText(child as Element);
-        else if (child.nodeType === Node.TEXT_NODE)
-          value += child.nodeValue || "";
-
-        // skip long text
-        if (value.length > 100) break;
-      }
-      if ((root as Element).shadowRoot)
-        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-        value += elementText((root as Element).shadowRoot!);
-    }
-  }
-
-  return value;
-}
-// --
-
 export const isVisible = (
   element: Element,
   computedStyle?: CSSStyleDeclaration
